feat: implement grocery list UI with shopping mode (#77)
Adds complete UI implementation for grocery lists feature:

**API Routes:**
- GET/POST /api/grocery-lists - List and create grocery lists
- GET/PATCH/DELETE /api/grocery-lists/[id] - List detail operations
- POST /api/grocery-lists/[id]/items - Add items to list
- PATCH/DELETE /api/grocery-lists/items/[itemId] - Toggle/update/delete items
- DELETE /api/grocery-lists/[id]/checked - Clear checked items
- POST /api/grocery-lists/[id]/share - Share list with users
- GET /api/grocery-lists/[id]/sync - Real-time sync polling

**Pages:**
- /grocery-lists - Main list view with filtering (all/my lists)
- /grocery-lists/[id] - List detail with items grouped by category
- /grocery-lists/[id]/shop - Full-screen shopping mode

**Features:**
- Optimistic UI updates for item toggling
- Wake Lock API to keep screen awake in shopping mode
- Real-time sync via 5-second polling
- Item grouping by category (produce, dairy, meat, etc.)
- Progress tracking with completion celebration
- "Clear checked" bulk action
- List sharing with view/edit permissions
- Shows who checked each item and when

**Components:**
- GroceryListClient, GroceryListDetailClient, ShoppingModeClient
- GroceryListCard, GroceryItem, GroceryListPageHeader
- CreateListModal, AddItemModal, ShareGroceryListModal

**Testing:**
- Unit tests for grocery-utils (18 tests)
- E2E tests for grocery lists flow (15 test cases)

+++ src/lib/grocery-db.ts
@@ -343,7 +343,7 @@ export async function toggleGroceryItem(
   const result = await query<GroceryListItemRow>(
     `UPDATE grocery_list_items
      SET checked = $2,
-         checked_by_user_id = CASE WHEN $2 = true THEN $3 ELSE NULL END,
+         checked_by_user_id = CASE WHEN $2 = true THEN $3::integer ELSE NULL END,
          checked_at = CASE WHEN $2 = true THEN NOW() ELSE NULL END,
          updated_at = NOW()
      WHERE id = $1
